requirements.txt
Django
django-jalali
djangorestframework
jdatetime
mssql-django
